Line 78 - Important note
Highlighted the important information in line 78. 

If an administrator is not aware of these changes they could manually migrate in PST files and leave the retention hold on for those mailboxes accidentally. 
This change provides clarity to the documentation and allows for the reader to skim the documentation and not miss the information that is buried in the list. 

Removed the question in the rest of the paragraph as well. The supporting information would be bulleted under the important box. 

Multiple administrators I have worked with did not know this information even though this is one of the only documents that provides information to import a PST file to Office 365.